refactor: introduce a client package
Added a new `client` package that will contain packages that may be used
by clients of flintlock (and flintlock itself). Initially this will be
**cloud-init** related functionality that is used by flintlock but it
will also be used by capmvm as it needs to augment the cloud-init
