+++ client/cloudinit/metadata.go
