+++ client/cloudinit/network.go
