+++ client/go.mod
@@ -1,0 +1,3 @@
+module github.com/weaveworks/flintlock/client
+
+go 1.17
